Release/2.5.0

+++ Sources/FastSocketKit/Model/Frame.swift
@@ -4,20 +4,21 @@
 //
 //  Created by Vinzenz Weist on 25.03.19.
 //  Copyright © 2019 Vinzenz Weist. All rights reserved.
+//
 import Foundation
 
-// 0                 1       N
-// +-----------------+-------+
-// |0|1 2 3 4 5 6 7 8|0 1 2 3|
-// +-+---------------+-------+
-// |O| FRAME LENGTH  |PAYLOAD|
-// |P|     (8)       |  (N)  |
-// |C|               |       |
-// +-+---------------+-------+
-// :Payload Data continued...:
-// + - - - - - - - - - - - - +
-// |Payload Data continued...|
-// +-------------------------+
+// 0         1           N
+// +---------+-----------+
+// |0|1 2 3 4|0 1 2 3... |
+// +-+-------+-----------+
+// |O| FRAME |  PAYLOAD  |
+// |P| LENGTH|    (N)    |
+// |C|  (4)  |           |
+// +-+-------+-----------+
+// :Payload continued... :
+// + - - - - - - - - - - +
+// |Payload continued... |
+// +---------------------+
 //
 // This describes the framing protocol.
 // - OPC:
@@ -27,7 +28,7 @@ import Foundation
 //      - 0x3: this is the fin byte, which is part of OPC but is on the first place in the protocol
 //      - 0x6 - 0xF: this bytes are reserved
 // - FRAME LENGTH:
-//      - this uses 8 bytes to store the entire frame size as a big endian uint64 value
+//      - this uses 8 bytes to store the entire frame size as a big endian uint32 value
 // - PAYLOAD:
 //      - continued payload data
 
@@ -37,10 +38,10 @@ import Foundation
 internal final class Frame: FrameProtocol {
     private var readBuffer = Data()
     /// private property to get parse the overhead size of a frame
-    private var contentSize: UInt64 {
+    private var contentSize: UInt32 {
         guard readBuffer.count >= Constant.overheadSize else { return .zero }
         let size = Data(readBuffer[1...Constant.overheadSize - 1])
-        return UInt64(size.integer)
+        return size.integer
     }
     /// crate instance of Frame
     internal required init() {
@@ -56,11 +57,11 @@ internal final class Frame: FrameProtocol {
         case let message as String:
             let message = message.utf8
             frame.append(Opcode.string.rawValue)
-            frame.append(UInt64(message.count + Constant.overheadSize).data)
+            frame.append(UInt32(message.count + Constant.overheadSize).data)
             frame.append(message)
         case let message as Data:
             frame.append(Opcode.data.rawValue)
-            frame.append(UInt64(message.count + Constant.overheadSize).data)
+            frame.append(UInt32(message.count + Constant.overheadSize).data)
             frame.append(message)
         default:
             throw FastSocketError.unknownOpcode
